+++ src/pages/StudyListPage/StudyComponents/Card.jsx
@@ -1,10 +1,9 @@
-import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import styles from './Card.module.css';
 import Emoji from '../../../components/Emoji/Emoji';
 import TotalPoint from '../../../components/TotalPoint/TotalPoint';
-import { getEmojis } from '../../../services/StudyDetailService';
 import { saveRecentStudy } from '../../../services/StudyService';
+import useTopEmojis from './useTopEmojis';
 import {
   getStudyBackgroundColor,
   getStudyBackgroundImage,
@@ -13,37 +12,7 @@ import {
 } from './studyBackground';
 
 const Card = ({ study }) => {
-  const [topEmojis, setTopEmojis] = useState([]);
-
-  useEffect(() => {
-    let isMounted = true;
-
-    const fetchTopEmojis = async () => {
-      try {
-        const emojis = await getEmojis(study.id);
-        const nextTopEmojis = (emojis ?? [])
-          .slice()
-          .sort((a, b) => (b.count ?? 0) - (a.count ?? 0))
-          .slice(0, 3);
-
-        if (isMounted) {
-          setTopEmojis(nextTopEmojis);
-        }
-      } catch (error) {
-        console.error(error);
-
-        if (isMounted) {
-          setTopEmojis([]);
-        }
-      }
-    };
-
-    fetchTopEmojis();
-
-    return () => {
-      isMounted = false;
-    };
-  }, [study.id]);
+  const topEmojis = useTopEmojis(study.id);
 
   const background = study.background;
   const hasImageBackground = isImageBackground(background);
@@ -83,34 +52,30 @@ const Card = ({ study }) => {
     >
       <article className={cardClassName} style={cardStyle}>
         <div className={styles.cardInner}>
-        {/* 카드 타이틀 + 포인트 */}
+          {/* 카드 타이틀 + 포인트 */}
           <div className={styles.titleDate}>
             <div className={styles.cardHeader}>
               <h3 className={styles.cardTitle}>
                 <span className={styles.nickname} style={nicknameStyle}>
                   {study.nickname}
-              </span>
+                </span>
                 의 {study.title}
               </h3>
-            {/* 획득 포인트 연결 */}
+              {/* 획득 포인트 연결 */}
               <TotalPoint id={study.id} theme={pointTheme} />
             </div>
 
-          {/* N일째 진행 중 */}
+            {/* N일째 진행 중 */}
             <p className={styles.cardProgressText}>{study.progressText}</p>
           </div>
 
-        {/* 스터디 설명 */}
+          {/* 스터디 설명 */}
           <p className={styles.cardDescription}>{study.description}</p>
 
-        {/* 이모지 */}
+          {/* 이모지 */}
           <div className={styles.cardEmojiTagDiv}>
             {topEmojis.map((emoji) => (
-              <Emoji
-                key={emoji.id}
-                emoji={emoji.emoji}
-                count={emoji.count}
-              />
+              <Emoji key={emoji.id} emoji={emoji.emoji} count={emoji.count} />
             ))}
           </div>
         </div>
